+++ tests_unit/mockgraphdatamodel.h
@@ -8,6 +8,6 @@ class MockGraphDataModel : public GraphDataModel {
     public:
         MockGraphDataModel() : GraphDataModel(new SettingsModel()) {}
 
-        MOCK_METHOD2(isPresent, bool(quint16 addr, quint16 bitmask));
+        MOCK_METHOD2(isPresent, bool(quint16 addr, quint32 bitmask));
 };
